Add FreeCanvas
+ GDI_FreeCanvas

+++ Runtime/Runtime/Random.cpp
@@ -4,7 +4,7 @@ namespace Sumomo2::Math {
 	float RandomRange(float min, float max)
 	{
 		float len = max - min;
-		return rand() % 10000 / 10000.0 * len + min;
+		return static_cast<float>(rand() % 10000 / 10000.0 * len + min);
 	}
 
 	void SetSeed() {
